Separate R & C++ files for SMLN families.
C++ file is now separate files for each SMLN family. The R user file has also been separated. The R internal file has not been seperated so it is easier to keep track of which functions still need to be converted.
