+++ R/Internal_Codes.R
@@ -180,7 +180,6 @@ MCMCR.sigma2.LN <- function(N, thin, Time, Cens, X, beta0, sigma20, logt0,
 
 ### OTHER FUNCTIONS REQUIRED FOR THE IMPLEMENTATION OF THE LOG-STUDENT'S T MODEL
 
-
 log.lik.LST <- function(Time, Cens, X, beta, sigma2, nu, set, eps_l, eps_r) {
     n <- length(Time)
     aux <- rep(0, n)
@@ -1323,106 +1322,6 @@ MCMCR.sigma2.LLOG <- function(N, thin, Q, Time, Cens, X, beta0, sigma20, logt0,
 
     chain <- cbind(beta, lambda, logt)
     return(chain)
-}
-
-
-#### Input check
-MCMC.param.check <- function(N, thin, burn, Time, Cens, X, beta0, sigma20,
-                             prior, set, eps_l, eps_r) {
-
-    num.obs <- nrow(X)
-    num.covariates <- ncol(X)
-
-    if (is.matrix(X) == F) {
-        stop("X is not a matrix.\n")
-    }
-
-    # Check N is a  0 < integer
-    if (N <= 0 | N %% 1 != 0) {
-        stop("N should be an integer greater than zero.\n")
-    }
-
-    if (thin < 2 | thin %% 1 != 0) {
-        stop("thin should be a integer > 2.\n")
-    }
-
-    if (burn < 0 | burn %% 1 != 0) {
-        stop("burn should be a non-negative integer.\n")
-    }
-
-    if (N < burn) {
-        stop("N must be greater than burn.\n")
-    }
-
-    if (burn %% thin != 0) {
-        stop("burn must a multiple of thin.\n")
-    }
-
-    if (N %% thin != 0) {
-        stop("N must a multiple of thin.\n")
-    }
-
-    if (all(Time > 0) == F) {
-        stop("All values in Time should be non-negative.\n")
-    }
-
-    if (length(Time) != num.obs) {
-    stop("Time is not the correct length.\n")
-    }
-
-    if ( all(Cens == 1 | Cens == 0) == F) {
-        stop("Cens should be either 0 or 1 for each observation\n")
-    }
-    if (length(Cens) != num.obs) {
-        stop("Cens is not the correct length.\n")
-    }
-
-    if (length(beta0) != num.covariates) {
-        stop("beta0 is not the correct length.\n")
-    }
-
-    if (prior %in% c(1, 2, 3) == F) {
-        stop("prior should be 1, 2 or 3. See documentation\n")
-    }
-
-    if (set %in% c(1, 2) == F) {
-        stop("set should be 1 or 2. See documentation\n")
-    }
-
-    if (burn == 0) {
-        cat(paste0("Note! No burn-in period is being used!\n"))
-    }
-
-}
-
-beta.sample <- function(n) {
-    cat("Sampling initial betas from a Normal(0, 1) distribution\n")
-    betas <- stats::rnorm(n, 0, 1)
-    cat(paste("Initial beta", 1:length(betas), ":", round(betas, 2), "\n"))
-    cat("\n")
-    return(betas)
-}
-
-sigma2.sample <- function() {
-    cat("Sampling initial sigma^2 from a Gamma(2, 2) distribution\n")
-    sigma2 <- stats::rgamma(1, 2, 2)
-    cat(paste("Initial sigma^2 :", round(sigma2, 2), "\n\n"))
-    return(sigma2)
-}
-
-nu.sample <- function() {
-    cat("Sampling initial nu from a Gamma(2, 2) distribution\n")
-    nu <- stats::rgamma(1, 2, 2)
-    cat(paste("Initial nu :", round(nu, 2), "\n\n"))
-    return(nu)
-
-}
-
-alpha.sample <- function() {
-    cat("Sampling initial alpha from a Uniform(1, 2) distribution\n")
-    alpha <- stats::runif(1, 1, 2)
-    cat(paste("Initial alpha :", round(alpha, 2), "\n\n"))
-    return(alpha)
 }
 
 pnormp <- function(q, mu = 0, sigmap = 1, p = 2,
