+++ R/RcppExports.R
@@ -1,72 +1,16 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-prior_LN <- function(beta, sigma2, prior, logs) {
-    .Call(`_BASSLINE_prior_LN`, beta, sigma2, prior, logs)
-}
-
-prior_nu_single <- function(nu, prior) {
-    .Call(`_BASSLINE_prior_nu_single`, nu, prior)
-}
-
-prior_LST <- function(beta, sigma2, nu, prior, logs) {
-    .Call(`_BASSLINE_prior_LST`, beta, sigma2, nu, prior, logs)
-}
-
-J_alpha <- function(alpha, k) {
-    .Call(`_BASSLINE_J_alpha`, alpha, k)
-}
-
-J_alpha_single <- function(alpha, k) {
-    .Call(`_BASSLINE_J_alpha_single`, alpha, k)
-}
-
-II_alpha <- function(alpha) {
-    .Call(`_BASSLINE_II_alpha`, alpha)
-}
-
-II_alpha_single <- function(alpha) {
-    .Call(`_BASSLINE_II_alpha_single`, alpha)
-}
-
-I_alpha <- function(alpha) {
-    .Call(`_BASSLINE_I_alpha`, alpha)
-}
-
-I_alpha_single <- function(alpha) {
-    .Call(`_BASSLINE_I_alpha_single`, alpha)
-}
-
-prior_alpha <- function(alpha, k, prior) {
-    .Call(`_BASSLINE_prior_alpha`, alpha, k, prior)
-}
-
-prior_alpha_single <- function(alpha, k, prior) {
-    .Call(`_BASSLINE_prior_alpha_single`, alpha, k, prior)
-}
-
 prior_LEP <- function(beta, sigma2, alpha, prior, logs) {
     .Call(`_BASSLINE_prior_LEP`, beta, sigma2, alpha, prior, logs)
-}
-
-r_GIG <- function(r) {
-    .Call(`_BASSLINE_r_GIG`, r)
 }
 
 d_texp <- function(x, trunc) {
     .Call(`_BASSLINE_d_texp`, x, trunc)
 }
 
-alpha_nu <- function(nu0, nu1, lambda, prior) {
-    .Call(`_BASSLINE_alpha_nu`, nu0, nu1, lambda, prior)
-}
-
 MH_marginal_sigma2 <- function(omega2, logt, X, beta, alpha, sigma20, prior) {
     .Call(`_BASSLINE_MH_marginal_sigma2`, omega2, logt, X, beta, alpha, sigma20, prior)
-}
-
-MH_nu_LST <- function(omega2, beta, lambda, nu0, prior) {
-    .Call(`_BASSLINE_MH_nu_LST`, omega2, beta, lambda, nu0, prior)
 }
 
 alpha_beta <- function(beta_0, beta_1, logt, X, sigma2, alpha) {
@@ -101,24 +45,56 @@ log_lik_LEP <- function(Time, Cens, X, beta, sigma2, alpha, set, eps_l, eps_r) {
     .Call(`_BASSLINE_log_lik_LEP`, Time, Cens, X, beta, sigma2, alpha, set, eps_l, eps_r)
 }
 
-RS_lambda_obs_LLOG <- function(logt, X, beta, sigma2, obs, N_AKS) {
-    .Call(`_BASSLINE_RS_lambda_obs_LLOG`, logt, X, beta, sigma2, obs, N_AKS)
-}
-
 Post_u_obs_LEP <- function(obs, ref, X, chain) {
     .Call(`_BASSLINE_Post_u_obs_LEP`, obs, ref, X, chain)
 }
 
-Post_lambda_obs_LST <- function(obs, ref, X, chain) {
-    .Call(`_BASSLINE_Post_lambda_obs_LST`, obs, ref, X, chain)
+r_GIG <- function(r) {
+    .Call(`_BASSLINE_r_GIG`, r)
+}
+
+RS_lambda_obs_LLOG <- function(logt, X, beta, sigma2, obs, N_AKS) {
+    .Call(`_BASSLINE_RS_lambda_obs_LLOG`, logt, X, beta, sigma2, obs, N_AKS)
+}
+
+prior_LN <- function(beta, sigma2, prior, logs) {
+    .Call(`_BASSLINE_prior_LN`, beta, sigma2, prior, logs)
+}
+
+J_alpha <- function(alpha, k) {
+    .Call(`_BASSLINE_J_alpha`, alpha, k)
+}
+
+J_alpha_single <- function(alpha, k) {
+    .Call(`_BASSLINE_J_alpha_single`, alpha, k)
+}
+
+II_alpha <- function(alpha) {
+    .Call(`_BASSLINE_II_alpha`, alpha)
+}
+
+II_alpha_single <- function(alpha) {
+    .Call(`_BASSLINE_II_alpha_single`, alpha)
+}
+
+I_alpha <- function(alpha) {
+    .Call(`_BASSLINE_I_alpha`, alpha)
+}
+
+I_alpha_single <- function(alpha) {
+    .Call(`_BASSLINE_I_alpha_single`, alpha)
+}
+
+prior_alpha <- function(alpha, k, prior) {
+    .Call(`_BASSLINE_prior_alpha`, alpha, k, prior)
+}
+
+prior_alpha_single <- function(alpha, k, prior) {
+    .Call(`_BASSLINE_prior_alpha_single`, alpha, k, prior)
 }
 
 rtnorm <- function(n, lower, upper, mu, sd) {
     .Call(`_BASSLINE_rtnorm`, n, lower, upper, mu, sd)
-}
-
-log_lik_LST <- function(Time, Cens, X, beta, sigma2, nu, set, eps_l, eps_r) {
-    .Call(`_BASSLINE_log_lik_LST`, Time, Cens, X, beta, sigma2, nu, set, eps_l, eps_r)
 }
 
 mvrnormArma <- function(n, mu, Sigma) {
@@ -135,5 +111,29 @@ logt_update_SMLN <- function(Time, Cens, X, beta, sigma2, set, eps_l, eps_r) {
 
 MCMC_LN_CPP <- function(N, thin, burn, Time, Cens, X, beta0, sigma20, prior, set, eps_l, eps_r) {
     .Call(`_BASSLINE_MCMC_LN_CPP`, N, thin, burn, Time, Cens, X, beta0, sigma20, prior, set, eps_l, eps_r)
+}
+
+prior_nu_single <- function(nu, prior) {
+    .Call(`_BASSLINE_prior_nu_single`, nu, prior)
+}
+
+prior_LST <- function(beta, sigma2, nu, prior, logs) {
+    .Call(`_BASSLINE_prior_LST`, beta, sigma2, nu, prior, logs)
+}
+
+alpha_nu <- function(nu0, nu1, lambda, prior) {
+    .Call(`_BASSLINE_alpha_nu`, nu0, nu1, lambda, prior)
+}
+
+MH_nu_LST <- function(omega2, beta, lambda, nu0, prior) {
+    .Call(`_BASSLINE_MH_nu_LST`, omega2, beta, lambda, nu0, prior)
+}
+
+log_lik_LST <- function(Time, Cens, X, beta, sigma2, nu, set, eps_l, eps_r) {
+    .Call(`_BASSLINE_log_lik_LST`, Time, Cens, X, beta, sigma2, nu, set, eps_l, eps_r)
+}
+
+Post_lambda_obs_LST <- function(obs, ref, X, chain) {
+    .Call(`_BASSLINE_Post_lambda_obs_LST`, obs, ref, X, chain)
 }
