+++ man/BASSLINE_convert.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/Utils.R
 \name{BASSLINE_convert}
 \alias{BASSLINE_convert}
 \title{Convert dataframe with mixed variables to a numeric matrix}
