+++ man/BF_lambda_obs_LLAP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLaplace.R
 \name{BF_lambda_obs_LLAP}
 \alias{BF_lambda_obs_LLAP}
 \title{Outlier detection for observation for the log-Laplace model}
