+++ man/BF_lambda_obs_LLOG.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLogistic.R
 \name{BF_lambda_obs_LLOG}
 \alias{BF_lambda_obs_LLOG}
 \title{Outlier detection for observation for the log-logisitc model}
