+++ man/BF_lambda_obs_LST.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogStudentsT.R
 \name{BF_lambda_obs_LST}
 \alias{BF_lambda_obs_LST}
 \title{Outlier detection for observation for the log-student's t model}
