+++ man/BF_u_obs_LEP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogExponentialPower.R
 \name{BF_u_obs_LEP}
 \alias{BF_u_obs_LEP}
 \title{Outlier detection for observation for the log-exponential power model}
