+++ man/CaseDeletion_LEP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogExponentialPower.R
 \name{CaseDeletion_LEP}
 \alias{CaseDeletion_LEP}
 \title{Case deletion analysis for the log-exponential power model}
