+++ man/CaseDeletion_LLOG.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLogistic.R
 \name{CaseDeletion_LLOG}
 \alias{CaseDeletion_LLOG}
 \title{Case deletion analysis for the log-logistic model}
