+++ man/CaseDeletion_LN.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogNormal.R
 \name{CaseDeletion_LN}
 \alias{CaseDeletion_LN}
 \title{Case deletion analysis for the log-normal model}
