+++ man/CaseDeletion_LST.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogStudentsT.R
 \name{CaseDeletion_LST}
 \alias{CaseDeletion_LST}
 \title{Case deletion analysis for the log-student's t model}
