+++ man/DIC_LEP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogExponentialPower.R
 \name{DIC_LEP}
 \alias{DIC_LEP}
 \title{Deviance information criterion for the log-exponential power model}
