+++ man/DIC_LLAP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLaplace.R
 \name{DIC_LLAP}
 \alias{DIC_LLAP}
 \title{Deviance information criterion for the log-Laplace model}
