+++ man/DIC_LLOG.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLogistic.R
 \name{DIC_LLOG}
 \alias{DIC_LLOG}
 \title{Deviance information criterion for the log-logistic model}
