+++ man/DIC_LN.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogNormal.R
 \name{DIC_LN}
 \alias{DIC_LN}
 \title{Deviance information criterion for the log-normal model}
