+++ man/LML_LEP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogExponentialPower.R
 \name{LML_LEP}
 \alias{LML_LEP}
 \title{Log-marginal likelihood estimator for the log-exponential power model}
