+++ man/LML_LLAP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLaplace.R
 \name{LML_LLAP}
 \alias{LML_LLAP}
 \title{Log-marginal likelihood estimator for the log-Laplace model}
