+++ man/LML_LLOG.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLogistic.R
 \name{LML_LLOG}
 \alias{LML_LLOG}
 \title{Log-marginal likelihood estimator for the log-logistic model}
