+++ man/LML_LN.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogNormal.R
 \name{LML_LN}
 \alias{LML_LN}
 \title{Log-marginal Likelihood estimator for the log-normal model}
