+++ man/MCMC_LEP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogExponentialPower.R
 \name{MCMC_LEP}
 \alias{MCMC_LEP}
 \title{MCMC algorithm for the log-exponential power model}
