+++ man/MCMC_LLAP.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogLaplace.R
 \name{MCMC_LLAP}
 \alias{MCMC_LLAP}
 \title{MCMC algorithm for the log-Laplace model}
