+++ man/MCMC_LN.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogNormal.R
 \name{MCMC_LN}
 \alias{MCMC_LN}
 \title{MCMC algorithm for the log-normal model}
