+++ man/MCMC_LST.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/LogStudentsT.R
 \name{MCMC_LST}
 \alias{MCMC_LST}
 \title{MCMC algorithm for the log-student's t model}
