+++ man/Trace_Plot.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/User_Codes.R
+% Please edit documentation in R/Utils.R
 \name{Trace_plot}
 \alias{Trace_plot}
 \title{Produce a trace plot of a variable's MCMC chain}
