+++ src/uses.h
@@ -1,0 +1,8 @@
+#define RCPP_ARMADILLO_RETURN_ANYVEC_AS_VECTOR
+#include <RcppArmadillo.h>
+
+using Rcpp::NumericVector;
+using Rcpp::NumericMatrix;
+using arma::vec;
+using Rcpp::List;
+using arma::mat;
